requirements.txt
fastapi==0.115.8
fastapi-cli==0.0.7
matplotlib==3.10.0
matplotlib-inline==0.1.7
mysql-connector-python==9.2.0
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.1.0
pydantic==2.10.6
pydantic-extra-types==2.10.2
pydantic-settings==2.7.1
pydantic_core==2.27.2
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
streamlit==1.43.0
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
uvicorn==0.34.0
